Remove one level of indirection brought by usecase
